% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_extract.R
\name{arch_extract}
\alias{arch_extract}
\title{Extract data from archive efficiently}
\usage{
arch_extract(
  police_force = ".",
  data_series = ".",
  month_min = ".",
  month_max = "."
)
}
\arguments{
\item{police_force}{A string.}

\item{data_series}{A string.}

\item{month_min}{A string.}

\item{month_max}{A string.}
}
\value{
A tibble.
}
\description{
Extract data from archive efficiently
}
\examples{
arch_extract(police_force = "west-midlands", data_series = "stop-and-search", month_min = "2015-01", month_max = "2020-08")
arch_extract(police_force = "greater-manchester", data_series = "outcomes", month_min = "2012-12", month_max = "2013-01")
}
