% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulating_arch_contents.R
\name{arch_filter}
\alias{arch_filter}
\title{Filters by force, data series, time frame}
\usage{
arch_filter(
  df = archive_contents,
  police_force = ".",
  data_series = ".",
  month_min = ".",
  month_max = ".",
  option
)
}
\arguments{
\item{df}{A tibble.}

\item{police_force}{A string.}

\item{data_series}{A string.}

\item{month_min}{A string.}

\item{month_max}{A string.}

\item{option}{Either contains or should_contain}
}
\value{
A tibble.
}
\description{
Filters by force, data series, time frame
}
\examples{
arch_filter(police_force = "wiltshire", data_series = "stop-and-search", month_min = "2019-01", month_max = "2024-06")
}
