% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_extract.R
\name{choose_files}
\alias{choose_files}
\title{For given force, data series, time range, chooses list of files to download}
\usage{
choose_files(
  police_force = ".",
  data_series = ".",
  month_min = ".",
  month_max = "."
)
}
\arguments{
\item{month_max}{}
}
\value{
List of urls to data files
}
\description{
For given force, data series, time range, chooses list of files to download
}
\examples{
choose_files(police_force = "dorset", data_series = "street", month_min = "2017-03", month_max = "2017-06")
}
