% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_extract.R
\name{extract_list}
\alias{extract_list}
\title{Takes a list of files, extracts them from archive}
\usage{
extract_list(url_list)
}
\arguments{
\item{url_list}{A list containing URLs of data files}
}
\value{
A list of data tables
}
\description{
Takes a list of files, extracts them from archive
}
