% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulating_arch_contents.R
\name{file_status}
\alias{file_status}
\title{Adds column detailing file status}
\usage{
file_status(df)
}
\arguments{
\item{df}{A tibble.}
}
\value{
A tibble.
}
\description{
Adds column detailing file status
}
