% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_arch_contents.R
\name{fs_graph}
\alias{fs_graph}
\title{Plots file presence/absence in folders}
\usage{
fs_graph(
  police_force = ".",
  data_series = ".",
  month_min = ".",
  month_max = "."
)
}
\arguments{
\item{police_force}{A string.}

\item{data_series}{A string.}

\item{month_min}{A string.}

\item{month_max}{A string.}
}
\value{
A ggplot object.
}
\description{
Plots file presence/absence in folders
}
\examples{
fs_graph(police_force = "wiltshire", data_series = "stop-and-search", month_min = "2019-01", month_max = "2024-06")
}
