% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulating_arch_contents.R
\name{should_contain}
\alias{should_contain}
\title{Adds column detailing whether folder should contain file}
\usage{
should_contain(df = archive_contents)
}
\arguments{
\item{df}{A tibble.}
}
\value{
A tibble.
}
\description{
Adds column detailing whether folder should contain file
}
