% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_extract.R
\name{arch_extract}
\alias{arch_extract}
\title{Extract data from archive efficiently}
\usage{
arch_extract(
  police_force = ".",
  data_series = ".",
  month_min = ".",
  month_max = "."
)
}
\arguments{
\item{police_force}{A string.}

\item{data_series}{A string.}

\item{month_min}{A string.}

\item{month_max}{A string.}
}
\value{
A tibble.
}
\description{
Extract data from archive efficiently
}
